Create file ddspan.go
